+++ backend/app/Models/User.php
@@ -7,6 +7,7 @@ use Filament\Models\Contracts\FilamentUser;
 use Filament\Panel;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -96,6 +97,16 @@ class User extends Authenticatable implements FilamentUser, HasMedia, MustVerify
     public function notifications(): HasMany
     {
         return $this->hasMany(Notification::class);
+    }
+
+    public function viewablePets(): BelongsToMany
+    {
+        return $this->belongsToMany(Pet::class, 'pet_viewers')->withTimestamps();
+    }
+
+    public function editablePets(): BelongsToMany
+    {
+        return $this->belongsToMany(Pet::class, 'pet_editors')->withTimestamps();
     }
 
     public function notificationPreferences(): HasMany